+++ rollup.config.js
@@ -5,6 +5,7 @@ import { uglify } from "rollup-plugin-uglify";
 import postcss from "rollup-plugin-postcss";
 import pkg from "./package.json";
 import json from "@rollup/plugin-json";
+import image from "rollup-plugin-img";
 
 export default {
   input: pkg.source,
@@ -14,6 +15,9 @@ export default {
   ],
   plugins: [
     json(),
+    image({
+      limit: 10000,
+    }),
     external(),
     babel({
       exclude: "node_modules/**",