+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Button test cases/stories/html (react.js)
 - Image Component
 - Link Component
+- Message Component
 - Header Component
 - Menu Component